backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.36
aiomysql==0.2.0
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
faster-whisper==1.0.3